use displayorder attribute to order clauses in HTML/DOC: https://github.com/metanorma/metanorma-ogc/issues/567
