+++ lib/isodoc/iec/base_convert.rb
@@ -15,35 +15,7 @@ module IsoDoc
         end
       end
 
-      def middle_title(_isoxml, out)
-        title1, title2 = middle_title_parts(out)
-        out.p(class: "zzSTDTitle1") do |p|
-          p.b { |b| b << title1 }
-        end
-        if title2
-          out.p(class: "zzSTDTitle1") { |p| p << "&#xa0;" }
-          out.p(class: "zzSTDTitle2") do |p|
-            p.b { |b| b << title2 }
-          end
-        end
-        out.p(class: "zzSTDTitle1") { |p| p << "&#xa0;" }
-      end
-
-      def middle_title_parts(_out)
-        title1 = @meta.get[:doctitlemain]&.sub(/\s+$/, "")
-        @meta.get[:doctitleintro] and
-          title1 = "#{@meta.get[:doctitleintro]} &#x2014; #{title1}"
-        title2 = nil
-        if @meta.get[:doctitlepart]
-          title1 += " &#x2014;"
-          title2 = @meta.get[:doctitlepart]&.sub(/\s+$/, "")
-          @meta.get[:doctitlepartlabel] and
-            title2 = "#{@meta.get[:doctitlepartlabel]}: #{title2}"
-        end
-        [title1, title2]
-      end
-
-      def bibliography(isoxml, out)
+      def bibliography(node, out)
         return super unless @is_iev
       end
